requirements.txt
streamlit>=1.29.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
lxml>=5.0.0
openai>=1.6.0
python-dotenv>=1.0.0
selenium>=4.15.0
webdriver-manager>=4.0.1